fix : fix Integer type to Long type

+++ backend/sswm/src/main/java/com/ground/sswm/auth/controller/AuthController.java
@@ -86,7 +86,7 @@ public class AuthController {
         throws InvalidTokenException {
         log.debug("[POST] /refresh-access-token " + refreshToken);
         Map<String, Object> claims = authService.getClaimsFromToken(refreshToken);
-        Auth saved = authService.getSavedTokenByUserId((int) claims.get("id"));
+        Auth saved = authService.getSavedTokenByUserId((Long) claims.get("id"));
         if (refreshToken.equals(saved.getRefreshToken())) {
             String accessToken = authService.createAccessToken(claims);
             saved.setAccessToken(accessToken);
